+++ src/drafts.cpp
@@ -113,11 +113,12 @@ void wps::check_duration( const uint8_t duration )
     check( duration <= 6, "[duration] must not exceed 6 monthly periods" );
 }
 
-void wps::check_monthly_budget( const eosio::asset monthly_budget)
+void wps::check_monthly_budget( const eosio::asset monthly_budget )
 {
     check( _settings.exists(), "settings must first be initialized");
     auto settings = _settings.get_or_default();
 
     check( monthly_budget.symbol == symbol{"EOS", 4}, "[monthly_budget] must use EOS symbol" );
     check( monthly_budget >= settings.deposit_required, "[monthly_budget] must be a minimum of " + settings.deposit_required.to_string());
+    check( monthly_budget <= settings.max_monthly_budget, "[monthly_budget] cannot be greater than [max_monthly_budget] of " + settings.max_monthly_budget.to_string());
 }