+++ src/vote.cpp
@@ -1,24 +1,27 @@
+// @action
 void wps::vote( const eosio::name voter, const eosio::name proposal_name, const eosio::name vote )
 {
     require_auth( voter );
     const eosio::name ram_payer = get_self();
 
+    // validation
     auto proposals_itr = _proposals.find( proposal_name.value );
-    check( proposals_itr != _proposals.end(), "[proposal_name] does not exists");
-
     auto votes_itr = _votes.find( proposal_name.value );
-    check( votes_itr != _votes.end(), "[proposal_name] does not exist");
-    check( vote == "yes"_n || vote == "no"_n || vote == "abstain"_n || vote == ""_n, "[vote] invalid (ex: yes/no/abstain)");
-    check( proposals_itr->end > current_time_point(), "proposal has ended");
 
-    _votes.modify( votes_itr, ram_payer, [&]( auto& row ) {
-        check(row.votes[voter] != vote, "[vote] has not been modified");
+    check( proposals_itr != _proposals.end(), "[proposal_name] does not exists");
+    check( proposals_itr->end > current_time_point(), "[proposal_name] has ended");
+    check( votes_itr != _votes.end(), "[proposal_name] votes does not exist");
+    check( vote == "yes"_n || vote == "no"_n || vote == "abstain"_n, "[vote] invalid (ex: yes/no/abstain)");
 
-        if (vote == ""_n) row.votes.erase(voter); // unvote
-        else row.votes[voter] = vote; // vote
+    // update vote
+    _votes.modify( votes_itr, ram_payer, [&]( auto& vote_row ) {
+        check(vote_row.votes[voter] != vote, "[vote] has not been modified");
+        vote_row.votes[voter] = vote;
 
         // re-caculate votes
-        row.total_net_votes = calculate_total_net_votes( row.votes );
+        _proposals.modify( proposals_itr, ram_payer, [&]( auto& proposals_row ) {
+            proposals_row.total_net_votes = calculate_total_net_votes( vote_row.votes );
+        });
     });
 }